EDGAR -> Check for presence of cannelloni and cangw during setup
