+++ opendut-edgar/src/setup/constants.rs
@@ -65,3 +65,5 @@ pub mod netbird {
         unpack_dir().map(|dir| dir.join("netbird"))
     }
 }
+
+pub const REQUIRED_COMMAND_LINE_PROGRAMS: [(&str, &str);3] = [("systemctl", "--version"), ("cannelloni", "-h"), ("cangw", "-s")];
